backend/services/auth-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-redis>=5.3.0
redis>=4.5
requests>=2.31
django-cors-headers>=4.0
